Add search
This adds a very simple search. It requires to have the following in your config.toml:

```
[outputs]
  home = ["HTML", "JSON", "XML"]
  page = ["HTML"]
```

(A pony-tutorial PR for that change is here: [ponylang/pony-tutorial#360](https://github.com/ponylang/pony-tutorial/pull/360) ).

the search results need to be styled properly and also some peview (maybe with highlighted query terms) should be shown.

If anybody has some good js or css advice, feel free to apply it.